requirements.txt
flask
pyyaml
celery
flask-cors
flask-login
flask-script
flask-migrate
flask-sqlalchemy
